% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\docType{data}
\name{add}
\alias{add}
\title{A family of functions to add HTML 5 elements.}
\format{An object of class \code{list} of length 127.}
\usage{
add
}
\description{
A family of functions to add HTML 5 elements.
}
\keyword{datasets}
