% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{add_div}
\alias{add_div}
\title{TITLE}
\format{HI format}
\usage{
add_div(html, into, ...)
}
\arguments{
\item{html}{An HTML object, e.g. output from create_html().}

\item{into}{characters; an identifier. It could be a tag name, a element id or a class name.}

\item{...}{Other contents / attibutes to be passed to the HTML 5 tags.}
}
\description{
DESCRIPTION
}
\details{
HTML 5 div element

DETAILS
}
\note{
NOTE
}
