% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{add_elements}
\alias{add_elements}
\title{Generic function to insert an element into a HTML document}
\usage{
add_elements(html, into, el)
}
\arguments{
\item{html}{An HTML object, e.g. output from create_html().}

\item{into}{characters; an identifier. It could be a tag name, an element id or a class name.}

\item{el}{An HTML tag object, see htmltools::tag for details.}
}
\description{
Generic function to insert an element into a HTML document
}
\note{
One should use the \link{add} family of functions for valid HTML5 tags.
}
\examples{
create_html() \%>\% 
  add_elements(into = "body", htmltools::tags$div("I am a div"))
}
