% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{add_family}
\alias{add_family}
\title{Helper functions to add element to HTML}
\arguments{
\item{my_html}{An HTML object, e.g. output from create_html().}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}

\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}
}
\description{
Helper functions to add element to HTML
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
