% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_extra.R
\name{add_google_style_button}
\alias{add_google_style_button}
\title{Google style button}
\usage{
add_google_style_button(my_html, material_id, ..., into = "body")
}
\arguments{
\item{my_html}{An HTML object; output from 'create_html'.}

\item{material_id}{Material id from google material design, e.g. "play_circle_outline". More other options, see \url{https://material.io/icons/}.}

\item{...}{Element contents and attributes; attrbutes must be named. See references for
HTML5-tags guides.}

\item{into}{characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Adds google material-design icon / button.
}
\note{
The default cursor over the icon is set to be a pointer. Use "style = 'cursor: xxx'" for other
options. A list of options can be found at
\url{https://www.w3schools.com/cssref/playit.asp?filename=playcss_cursor&preval=pointer}.
}
