% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_extra.R
\name{add_google_toggle_button}
\alias{add_google_toggle_button}
\title{Add a toggle button with google material design}
\usage{
add_google_toggle_button(my_html, into, widget_id,
  material_1 = "play_circle_outline", material_2 = "pause_circle_outline",
  ...)
}
\description{
Add a toggle button with google material design
}
\note{
The two material-icons should have the same size.
}
\examples{
create_html() \%>\% 
  add_google_style() \%>\% 
  add_google_toggle_button(into = "body", widget_id = "gbutton_1") \%>\% 
  start_app()
}
