% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_extra.R
\name{add_grid}
\alias{add_grid}
\title{Add a grid of div's to the HTML}
\usage{
add_grid(my_html, nr, nc, first = "row", into = "body")
}
\arguments{
\item{my_html}{An HTML object; output from 'create_html'.}

\item{nr}{integer; number of rows to add.}

\item{nc}{integer; number of columns to add.}

\item{first}{"row" or "column"; if "row" first, then create `nr` rows, then in each row
make `nc` columns, and vice versa.}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Add a grid of div's to the HTML
}
\examples{
create_html() \%>\% add_grid(3, 4, first = "row")
}
