% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{add_js_library}
\alias{add_js_library}
\title{Add JavaScript libraries to a HTML file}
\usage{
add_js_library(my_html, js_libs)
}
\arguments{
\item{my_html}{An HTML object, e.g. output from create_html().}

\item{js_libs}{A character vector. The JavaScript libraries to use. Currently support 'plotly', 'p5', 'd3' and 'vega'.}
}
\description{
Add JS libraries links to html header
}
