% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_extra.R
\name{add_radio_button_group}
\alias{add_radio_button_group}
\title{Radio button group}
\usage{
add_radio_button_group(my_html, into, widget_id, values, labels,
  stack = T, ...)
}
\arguments{
\item{my_html}{An HTML object; output from 'create_html'.}

\item{into}{characters; an identifier. It could be a tag name, an element id or a class name.}

\item{widget_id}{character; an unique identifier for the widget.}

\item{values}{vector; the underlying values of the selections.}

\item{labels}{character vector; the labels shown in the app.}

\item{stack}{T or F; if TRUE, the radio buttons are stacked vertically.}

\item{...}{Element contents and attributes for the radio widget; attrbutes must be named.
See references for HTML5-tags guides.}
}
\description{
Radio button group
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
