% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/html_builder.R
\name{add_script}
\alias{add_script}
\alias{add_script}
\alias{add_script_from_link}
\alias{add_script_from_file}
\alias{add_js_library}
\title{Helper functions to add element to HTML}
\usage{
add_script(...)

add_script_from_link(...)

add_script_from_file(my_html, src, ...)

add_js_library(my_html, js_libs)
}
\arguments{
\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}

\item{my_html}{An HTML object, e.g. output from create_html().}

\item{src}{character; path to the local JS file.}

\item{js_libs}{A character vector. The JavaScript libraries to use. Currently support 'plotly', 'p5', 'd3' and 'vega'.}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Helper functions to add element to HTML

Add JS libraries links to html header
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
