% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_extra.R
\name{add_slider_with_text}
\alias{add_slider_with_text}
\title{Slider widget
A slider widget consists of 2 labels and a slider (<span> + <label> + <input>)}
\usage{
add_slider_with_text(my_html, into = "body", widget_id, label = "My Slider",
  label_id, value = 50, slider_id, ...)
}
\description{
Slider widget
A slider widget consists of 2 labels and a slider (<span> + <label> + <input>)
}
\examples{
create_html() \%>\% 
  add_slider(widget_id = "slider_1") \%>\% 
  start_app()
}
