% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/html_builder.R
\name{add_svgs}
\alias{add_svgs}
\alias{add_svg}
\alias{add_image}
\title{Helper functions to add element to HTML}
\usage{
add_svg(my_html, ..., into = "body")

add_image(my_html, ..., into = "body")
}
\arguments{
\item{my_html}{An HTML object, e.g. output from create_html().}

\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Helper functions to add element to HTML

Add svg element to the HTML object.

Add image SVG element to the HTML object.
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
