% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/html_builder.R
\name{add_text}
\alias{add_text}
\alias{add_text}
\alias{add_title}
\title{Helper functions to add element to HTML}
\usage{
add_text(...)

add_title(my_html, ..., size = 3, into = "body")
}
\arguments{
\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}

\item{my_html}{An HTML object, e.g. output from create_html().}

\item{size}{integer from 1 to 6; size of the title.}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Helper functions to add element to HTML
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
