% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/html_builder.R
\name{add_widget}
\alias{add_widget}
\alias{add_button}
\alias{add_slider}
\alias{add_radio_button}
\title{Helper functions to add element to HTML}
\usage{
add_button(...)

add_slider(...)

add_radio_button(...)
}
\arguments{
\item{...}{Takes three arguments.

\code{my_html} An HTML object, e.g. output from create_html().

\code{into} Characters; an identifier. It could be a tag name, an element id or a class name. 

\code{...} Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}
}
\description{
A widget is a <input> tag most of the time. Currently support
\code{add_button}, \code{add_slider} and \code{add_radio_button}. 
Other types can be added via the general \code{add$input(type = "XXX")} syntax.
Compound widgets can be formed using primitive widgets, see for example 
\link{add_google_style_button}, \link{add_google_toggle_button}, 
\link{add_counter}, \link{add_slider_with_text} and \link{add_radio_button_group}.
}
\examples{
\dontrun{
create_html() \%>\% 
  add_title("Button demo") \%>\% 
  add_button(value = "A Button") \%>\% 
  start_app()

create_html() \%>\% 
  add_title("Slider demo") \%>\% 
  add_slider() \%>\% 
  start_app()

create_html() \%>\% 
  add_title("Radio button demo") \%>\% 
  add_radio_button("A option") \%>\% 
  start_app()
}
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
