% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/html_builder.R
\name{add_widgets}
\alias{add_widgets}
\alias{add_button}
\alias{add_slider}
\alias{add_radio_button}
\title{Helper functions to add element to HTML}
\usage{
add_button(...)

add_slider(...)

add_radio_button(...)
}
\arguments{
\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}

\item{my_html}{An HTML object, e.g. output from create_html().}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Helper functions to add element to HTML
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
