% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_helpers.R
\name{builder}
\alias{builder}
\alias{add}
\title{A family of functions for adding HTML 5 elements.}
\usage{
add
}
\arguments{
\item{html}{An HTML object, e.g. output from create_html().}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}

\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}
}
\description{
The \code{add} environment wraps around the \code{htmltools::tags} environments, and it contains 
functions for all valid HTML5 tags.
}
\examples{
\dontrun{
library(magrittr)
library(jsvis)

create_html() \%>\% 
  add$title("A demo webpage", size = 2) \%>\% 
  add$div("I am a DIV") \%>\% 
  start_app()
}
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
