% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_start.R
\name{copy_assets}
\alias{copy_assets}
\title{Copy assets to target directory}
\usage{
copy_assets(path, target_dir, message = T)
}
\arguments{
\item{path}{character_string; path to the source file / folder.}

\item{target_dir}{character_string; path to the target folder.}

\item{message}{T or F; if T, report success.}
}
\description{
This is needed when one needs to serve local file to RStudio viewer.
}
\keyword{internal}
