% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_daemonise.R
\name{daemon_app}
\alias{daemon_app}
\title{Start and Stop daemonized app}
\description{
Start and Stop daemonized app
}
\section{Usage}{

\preformatted{
app <- daemon_app$new()

app$start_daemon_app(filename = NULL)

app$stop_daemon_app()
}
}

\section{Arguments}{

\code{filename} character string; the path to the app file.
}

\section{Methods}{

\code{$new()} starts a new process.

\code{$start_daemon_app()} takes a rjs app filepath argument and starts a
child process in the background.

\code{$stop_daemon_app()} stops the process.
}

