% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_id}
\alias{make_id}
\title{A help function to make multiple ids.}
\usage{
make_id(...)
}
\arguments{
\item{...}{Arguments to pass to \code{paste} and joined by '_'.}
}
\description{
This function is basically an alias of \code{paste}.
}
\examples{
make_id("item", 1:5)
# [1] "item_1" "item_2" "item_3" "item_4" "item_5"
}
