% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builder.R
\name{map_add}
\alias{map_add}
\title{Add a list of div's to the HTML}
\usage{
map_add(my_html, add_fun, n, id_prefix = "item_", into = "body")
}
\arguments{
\item{my_html}{An HTML object; output from 'create_html'.}

\item{add_fun}{function; the add function to be called}

\item{n}{integer; number of elements to add.}

\item{id_prefix}{character; prefix to be added to}

\item{into}{characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Add a list of div's to the HTML
}
\examples{
create_html() \%>\% map_add(add_row, 5, id_prefix = "row_")
}
