% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{preview_app}
\alias{preview_app}
\title{Preview app (Deprecated)}
\usage{
preview_app(...)
}
\arguments{
\item{...}{Arguments to pass to `start_app'. See help("start_app").}
}
\description{
This function is deprecated. Use \link{start_app} instead.
}
