% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builder.R
\name{set_default}
\alias{set_default}
\title{Set default of a function}
\usage{
set_default(FUN, ...)
}
\arguments{
\item{FUN}{A function.}

\item{...}{Named arguments.}
}
\description{
Sets default parameter of a function.
}
\examples{
test_fun <- function(a, b) { paste(a, b) }
test_fun_2 <- set_default(test_fun, a = "Hello")

test_fun("Hello", "World")
# "Hello World"
test_fun_2("World")
# "Hello World"
test_fun_2("World", a = "The")  # Can still override default if needed.
# "The World"
}
