% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_app_from_file}
\alias{start_app_from_file}
\title{Run an interactive app}
\usage{
start_app_from_file(filepath, user_function, server = F, assets_folder,
  host = "localhost", port = 9454, browser = "viewer", quiet = T)
}
\arguments{
\item{filepath}{Filepath to a HTML file.}

\item{user_function}{R function; the function to process the data from the web interface.}

\item{server}{T or F; whether to enable interaction between JS and R.}

\item{assets_folder}{path of the assets.}

\item{host}{character string; Address to host the app.}

\item{port}{integer; Port to host the app.}

\item{browser}{"browser" (web) or "viewer" (R).}

\item{quiet}{T or F; if T, display is suppressed.}
}
\description{
Run an interactive app
}
