% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_daemonised.R
\name{start_daemon_app}
\alias{start_daemon_app}
\title{Start daemonized app}
\usage{
start_daemon_app(filename)
}
\arguments{
\item{filename}{filepath to the app R file.}
}
\description{
Start daemonized app
}
