% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_step_through.R
\name{step_through}
\alias{step_through}
\title{Step through the building process of a html.}
\usage{
step_through(fun, start_ind = 1, sleep = 1)
}
\arguments{
\item{fun}{A function that consists of solely `add-*` functions.}

\item{start_ind}{An integer; the index at which to start the inspection.}

\item{sleep}{A positive number; the number of second to wait between the steps.}
}
\description{
Step through the building process of a html.
}
\examples{
\dontrun{
build0 <- . \%>\%
  add_title("Send message") \%>\%
  add_slider(min = "0", max = "100", oninput = "show_value(value)") \%>\%
  add_title("Receive message", size = 3) \%>\%
  add_row("I am a row", style = 'background-color: lightblue')
step_through(build0)
}
}
