% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_daemonised.R
\name{stop_daemon_app}
\alias{stop_daemon_app}
\title{Stop daemonized app}
\usage{
stop_daemon_app(handle)
}
\arguments{
\item{handle}{A handle returned by `subprocess::spawn_process`.}
}
\description{
Stop daemonized app
}
