% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/html_builder.R
\name{add_containers}
\alias{add_containers}
\alias{add_span}
\alias{add_div}
\alias{add_row}
\alias{add_column}
\alias{add_item}
\alias{add_container}
\title{Helper functions to add element to HTML}
\usage{
add_span(...)

add_div(...)

add_row(...)

add_column(...)

add_item(...)

add_container(...)
}
\arguments{
\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}

\item{my_html}{An HTML object, e.g. output from create_html().}

\item{into}{Characters; an identifier. It could be a tag name, an element id or a class name.}
}
\description{
Helper functions to add element to HTML
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
