% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{add_family}
\alias{add_family}
\title{Helper functions to add element to HTML}
\arguments{
\item{...}{Takes three arguments: 

\code{my_html} An HTML object, e.g. output from create_html(). 

\code{into} Characters; an identifier. It could be a tag name, an element id or a class name.  

\code{...} Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}
}
\description{
The add_* functions are divided into 6 categories. See 
\link{add_containers}, \link{add_text}, \link{add_widget}, \link{add_script}, 
\link{add_style} and \link{add_svg}.
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
