% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/html_builder.R
\name{add_script}
\alias{add_script}
\alias{add_script}
\alias{add_script_from_link}
\alias{add_script_from_file}
\alias{add_js_library}
\title{Helper functions to add element to HTML}
\usage{
add_script(...)

add_script_from_link(...)

add_script_from_file(my_html, src, inline = T,
  mime = "application/javascript", ...)

add_js_library(my_html, js_libs)
}
\arguments{
\item{...}{Takes three arguments for \code{add_script} and \code{add_script_from_link}.

\code{my_html} An HTML object, e.g. output from create_html().

\code{into} Characters; an identifier. It could be a tag name, an element id or a class name. 

\code{...} Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}

\item{my_html}{An HTML object, e.g. output from create_html().}

\item{src}{character; path to the local JS file.}

\item{inline}{T or F; if T, the entire content of the file is copied to the HTML.}

\item{mime}{Used if inline = F. The mime type of the UCI data. See reference for more detail.}

\item{js_libs}{A character vector. The JavaScript libraries to use. Currently support 'plotly', 'p5', 'd3' and 'vega'.}
}
\description{
Four ways to add script to HTML: \code{add_script}, \code{add_script_from_link}, 
\code{add_script_from_file} and \code{add_js_library}.

Adds inline script to HTML.

Adds a script from an external link.

Adds a script from local file; this is done inline or via data UCI.

Adds commonly used javascript liberaries via external links. 
A note would be made when this function is called.
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}

MIME types \url{https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types}
}
