% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builder.R
\name{builder}
\alias{builder}
\alias{add}
\title{A family of functions for adding HTML 5 elements.}
\usage{
add
}
\arguments{
\item{...}{Element contents and attributes; attrbutes must be named. See references for 
HTML5-tags guides.}
}
\description{
The \code{add} environment wraps around the \code{htmltools::tags} environments, and it contains 
functions for all valid HTML5 tags.
}
\examples{
\dontrun{
library(jsvis)
create_html() \%>\% 
  add_title("A demo webpage", size = 2) \%>\% 
  add$div("body", "I am a DIV") \%>\% 
  add_row("I am a DIV") \%>\% 
  start_app()
}
}
\references{
Quick guide to HTML tags \url{https://www.nobledesktop.com/html-quick-guide/}
}
