% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_daemonise.R
\docType{data}
\name{daemon_app}
\alias{daemon_app}
\title{Start and Stop daemonized app}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
\preformatted{
p <- daemon_app$new()

p$start_daemon_app(filename)

p$stop_daemon_app()
}
}
\arguments{
\item{filename}{filepath to the app R file.
Stop daemonized app}
}
\description{
Start and Stop daemonized app
}
\details{
\code{$new()} starts a new process.

\code{$start_daemon_app(filename)} takes a rjs app filepath argument.

\code{$stop_daemon_app()} stops the process.
A handle returned by `subprocess::spawn_process`.
Start daemonized app
}
\keyword{datasets}
