% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builder.R
\name{map_add}
\alias{map_add}
\title{Add a list of div's to the HTML}
\usage{
map_add(my_html, add_fun, args_list, n, into = "body", ...)
}
\arguments{
\item{my_html}{An HTML object; output from 'create_html'.}

\item{add_fun}{function; the add-html5-tag function to be called.}

\item{args_list}{a list of lists containing the arguments to pass to \code{add_fun}.}

\item{n}{integer; number of elements to add.}

\item{into}{characters; an identifier. It could be a tag name, an element id or a class name.}

\item{...}{(Optional) Other parameters to pass to \code{add_fun}.}
}
\description{
Add a list of div's to the HTML
}
\examples{
create_html() \%>\% map_add(add_row, list(id = make_id("row", 1:5)))
}
