% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_start.R
\name{start_app}
\alias{start_app}
\title{Run an interactive app}
\usage{
start_app(my_html, user_function = identity, server = F, assets,
  host = "localhost", port = 9454, browser = "viewer", quiet = F)
}
\arguments{
\item{my_html}{An HTML object, e.g. output from 'create_html', or a character 
string specifying a filepath to a HTML file.}

\item{user_function}{R function; the function to process the data from the web interface.}

\item{server}{T or F; whether to enable interaction between JS and R.}

\item{assets}{path of the assets. Note that it only applies when `server = F`.}

\item{host}{character string; Address to host the app.}

\item{port}{integer; Port to host the app.}

\item{browser}{"browser" (web) or "viewer" (R).}

\item{quiet}{T or F; if T, display is suppressed. This is useful when one 
runs `daemon_app`.}
}
\description{
Run an interactive app
}
\note{
The `assets' argument does not apply when server = T, because the server creates
its own temporary folder and serve the website from there. At the moment, it is unclear 
how to gain access to that folder, so it is recommended to use DataUCI instead to serve
local file for dynamic site.
}
